Mods to downloads directory to support more download clients
Create but don't start transmission. Updated settings.json

Not sure if the settings.json file gets created when the container is created. If yes, bonus, let's update it with customized settings. If not, maybe start then stop the container?
Corrected transmission folders


Fixed unable to connect to sabnzbd without a defined hostname