Python 3, Ansible Roles, revise package install step
By default, Ubuntu 18.04 LTS comes with python 3 only. Running this playbook from a remote system caused errors if not setting things to Python 3. Ironically, samba still needs python 2, so it gets installed there.

Updated the ansible roles in requirements.yml to the latest

The {{item}} method for the apt module is going to be deprecated (or so the warning screen told me). This is the now the approved away of installing multiple packages