Secured n8n
Put it behind a conditional. Now if you decide you want it secured
you can have it secured, but if you decide to expose it externally
it /must/ be secured. Now of course that must is a little bit
undermined by setting a default user and password. But it is at
least something.