add variable to match name used in task
In the transmission_with_openvpn task, the variable used is:
    `transmission_openvpn_available_externally`
In the transmission_with_openvpn defaults, the variable is:
    `transmission_with_openvpn_available_externally`
To reconcile this, I added the former variable into the default.
I left the extra `*_available_externally` variable in defaults in case it's
used elsewhere, but a quick search in the repo showed it's not used.
The extra one could be deleted, but I'll leave that up to the merger.