Try to fix drone CI gitea missing variable
From Github:

```
TASK [drone-ci : Add webhook allowed hosts to Gitea] ***************************
fatal: [instance]: FAILED! =>
The task includes an option with an undefined variable. The error was: 'gitea_data_directory' is undefined. 'gitea_data_directory' is undefined

The error appears to be in '/github/workspace/roles/drone-ci/tasks/main.yml': line 67, column 7, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

    - name: Add webhook allowed hosts to Gitea
      ^ here
```

I tried to find the nearest place these variables are supposed to be
defined for drone ci. We'll see how it goes.